Fix copyright notice performance when adding/removing items
This change avoids rescanning the list of map items every time an item
is added or removed, to recompute the Z value at which the copyright
notice should go.

Task-number: QTBUG-64880

+++ src/location/declarativemaps/qdeclarativegeomapcopyrightsnotice.cpp
@@ -297,7 +297,7 @@ bool QDeclarativeGeoMapCopyrightNotice::copyrightsVisible() const
 /*!
     \internal
 */
-void QDeclarativeGeoMapCopyrightNotice::setCopyrightsZ(int copyrightsZ)
+void QDeclarativeGeoMapCopyrightNotice::setCopyrightsZ(qreal copyrightsZ)
 {
     setZ(copyrightsZ);
     update();
